Split AttackComponent into specialized types
- Split AttackComponent into base class with shared targeting logic
- Add MeleeAttackComponent for direct damage/knockback attacks
- Add ProjectileAttackComponent for projectile-based attacks
- Update Bear to use MeleeAttackComponent as requested
- Update Turret to use ProjectileAttackComponent
- Rename performAttack → tryAttack and executeAttack → performAttack
- Align knockback implementation using applyImpulse method
